refactor(reouter): move search router to feature-router lib
