+++ libs/feature/router/src/index.ts
@@ -1,1 +1,2 @@
 export * from './lib/feature-router.module'
+export * from './lib/default'
