+++ libs/feature/search/src/index.ts
@@ -4,4 +4,3 @@ export * from './lib/state/selectors'
 export * from './lib/state/search.facade'
 export * from './lib/state/effects'
 export * from './lib/state/reducer'
-export * from './lib/router'
